Added folder for EDA (Exploratory Data Analysis).
Concatenates CSV files together, then uploads them to MongoDB in
batches. This allows all the logs of a certain module to be retrieved
quickly. An index is applied to the "course" field

Takes approximately two hours to run this program

An example of retrieving and processing a module is provided