Add lookup to Mongo database
Lookup between course_id (such as "CA169") and moodle_id (such as 31103)
added into Mongo database.

Note that a course_id can correspond to multiple moodle_ids. An example
of this is introduction to law, where the module is run in each semester
in different moodle_ids. An example of looking up a course_id with
multiple moodle_ids is provided