Bugfix: MongoDB dropping random logs
After switching to Mongo, I noticed from the graphs of Moodle accesses
that a lot of data appeared to be missing in random places. About 20% of
the data turned out to be missing compared to the pre-Mongo version

It turns out that I was processing the logs in batches of 100,000 in unsafe mode
(because I didn't want errors thrown when duplicates were entered), but
what actually happened was that whenever a duplicate was discovered the
rest of the batch was discarded

Switched from insert to save command, and not using batch mode anymore.

Fixes #1